Move tests.py to tests/test_base.py so pytest will find and run them
Seems we've been ignoring these tests for quite a while.
Also move tests in l10n_utils/tests/__init__.py so pytest can find them.
